Moved mocks and tests folder into the client folder. Pathed the jest configs to there also. Added AvgRating to the Cart component
