+++ client/__mocks__/fileMock.js
