+++ client/src/components/App.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import axios from 'axios';
 import Atelier from '../Atelier.js';
-import Overview from './overview/Overview.jsx';
+import Overview from './Overview.jsx';
 import Ratings from './RatingsReview.jsx';
 import RICWidget from './RIC-Widget.jsx';
 import AvgRating from './AvgRating.jsx';
